requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
openai
tiktoken
jinja2
pyyaml
